# packages/contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: Core Accounting MVP API
  version: 1.2.0
  description: 'مواصفات REST API للنواة المحاسبية القابلة للتوسع.


    مصادر التصميم: ERD v1.2، قاموس البيانات v1.2، مصفوفة الصلاحيات وقواعد الترحيل v1.1، وحالات الاستخدام وتدفقات العمل v1.0، مع تعديلات ADR-002 v1.3.


    قرارات تنفيذية مثبتة في هذه النسخة:

    - الشركة الحالية تُثبت في الجلسة عبر /auth/context، ولا يُعتمد companyId الوارد من العميل في العمليات التشغيلية.

    - رقم المستند يُحجز عند إنشاء المسودة مع قبول فجوات الأرقام الموثقة.

    - منشئ القيد اليدوي لا يرحله افتراضياً، ويمكن جعل ذلك إعداداً للشركات الصغيرة.

    - واجهة التخصيصات مدعومة في API، ولو أُجلت واجهتها في الإصدار الأول.

    - مدة الاحتفاظ بمفتاح Idempotency هي 24 ساعة افتراضياً وقابلة للضبط.

    - أقصى نطاق تفاعلي للتقارير 366 يوماً.


    المبالغ وأسعار الصرف والمعرفات الكبيرة تُنقل كنصوص للحفاظ على الدقة. الحقول المشفرة وBlind Indexes لا تظهر في الاستجابات
    العامة.


    تغييرات v1.1:

    - اعتماد counter_account_id رسمياً في سندات القبض والصرف العامة.

    - إضافة GET /dashboard/summary ببيانات مشتقة وصلاحيات حقلية.

    - لا تُخزن مؤشرات لوحة المتابعة كمصدر حقيقة مستقل.'
  contact:
    name: Core Accounting MVP Technical Team
  license:
    name: Internal Use
    identifier: LicenseRef-Internal
servers:
- url: /api/v1
  description: مسار نسبي يعمل على نفس أصل الواجهة في جميع البيئات
...
security:
- SessionCookie: []
paths:
...
  /sales-invoices:
    get:
      tags: [Sales Invoices]
      summary: عرض فواتير المبيعات والإشعارات الدائنة
      operationId: listSalesInvoices
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      - { name: documentType, in: query, schema: { type: string, enum: [SALES_INVOICE, SALES_CREDIT_NOTE] } }
      - { name: status, in: query, schema: { $ref: '#/components/schemas/AccountingDocumentStatus' } }
      - { name: customerId, in: query, schema: { type: string, pattern: '^[1-9][0-9]*$' } }
      - { name: outstandingOnly, in: query, schema: { type: boolean } }
      responses:
        '200': { description: قائمة النتائج, content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/SalesInvoice' } }, meta: { $ref: '#/components/schemas/PaginationMeta' } } } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
      x-permission: sales_invoices.view
      security: [{ SessionCookie: [] }]
    post:
      tags: [Sales Invoices]
      summary: إنشاء مسودة فاتورة أو إشعار دائن
      operationId: createSalesInvoice
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SalesInvoiceCreateRequest' } } } }
      responses:
        '201': { description: تم الإنشاء, content: { application/json: { schema: { $ref: '#/components/schemas/SalesInvoice' } } } }
        '422': { $ref: '#/components/responses/BusinessRuleViolation' }
      x-permission: sales_invoices.create
      security: [{ SessionCookie: [], CsrfToken: [] }]
  /sales-invoices/{invoiceId}:
    parameters:
    - { name: invoiceId, in: path, required: true, schema: { type: string, pattern: '^[1-9][0-9]*$' } }
    get:
      tags: [Sales Invoices]
      summary: عرض فاتورة أو إشعار دائن
      operationId: getSalesInvoice
      responses:
        '200': { description: نجاح, content: { application/json: { schema: { $ref: '#/components/schemas/SalesInvoice' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
      x-permission: sales_invoices.view
      security: [{ SessionCookie: [] }]
    patch:
      tags: [Sales Invoices]
      summary: تعديل مسودة فاتورة أو إشعار دائن
      operationId: updateSalesInvoice
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SalesInvoiceUpdateRequest' } } } }
      responses:
        '200': { description: نجاح, content: { application/json: { schema: { $ref: '#/components/schemas/SalesInvoice' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
      x-permission: sales_invoices.update
      security: [{ SessionCookie: [], CsrfToken: [] }]
  /sales-invoices/{invoiceId}/post:
    post:
      tags: [Sales Invoices]
      summary: ترحيل الفاتورة وتوليد القيد المتوازن
      operationId: postSalesInvoice
      parameters:
      - { name: invoiceId, in: path, required: true, schema: { type: string, pattern: '^[1-9][0-9]*$' } }
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [version], properties: { version: { type: integer, minimum: 0 } } } } } }
      responses:
        '200': { description: تم الترحيل }
        '400': { $ref: '#/components/responses/ValidationError' }
      x-permission: sales_invoices.post
      security: [{ SessionCookie: [], CsrfToken: [] }]
  /sales-invoices/{invoiceId}/cancel:
    post:
      tags: [Sales Invoices]
      summary: إلغاء مسودة فاتورة
      operationId: cancelSalesInvoice
      parameters:
      - { name: invoiceId, in: path, required: true, schema: { type: string, pattern: '^[1-9][0-9]*$' } }
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [version, reason], properties: { version: { type: integer, minimum: 0 }, reason: { type: string, minLength: 3 } } } } } }
      responses:
        '200': { description: تم الإلغاء }
        '400': { $ref: '#/components/responses/ValidationError' }
      x-permission: sales_invoices.cancel
      security: [{ SessionCookie: [], CsrfToken: [] }]
  /sales-invoices/{invoiceId}/reverse:
    post:
      tags: [Sales Invoices]
      summary: عكس فاتورة مرحلة غير مرتبطة بتحصيل
      operationId: reverseSalesInvoice
      parameters:
      - { name: invoiceId, in: path, required: true, schema: { type: string, pattern: '^[1-9][0-9]*$' } }
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [version, reversalDate, reason], properties: { version: { type: integer, minimum: 0 }, reversalDate: { type: string, format: date }, reason: { type: string, minLength: 3 } } } } } }
      responses:
        '200': { description: تم العكس }
        '422': { $ref: '#/components/responses/BusinessRuleViolation' }
      x-permission: sales_invoices.reverse
      security: [{ SessionCookie: [], CsrfToken: [] }]
...
  /tax-rates:
    get:
      tags: [Sales Invoices]
      summary: عرض نسب ضريبة المبيعات
      operationId: listTaxRates
      responses:
        '200': { description: نجاح, content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/TaxRate' } } } } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
      x-permission: sales_invoices.view
      security: [{ SessionCookie: [] }]
    post:
      tags: [Sales Invoices]
      summary: إنشاء نسبة ضريبة
      operationId: createTaxRate
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TaxRateInput' } } } }
      responses:
        '201': { description: تم الإنشاء, content: { application/json: { schema: { $ref: '#/components/schemas/TaxRate' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
      x-permission: tax_rates.manage
      security: [{ SessionCookie: [], CsrfToken: [] }]
  /tax-rates/{taxRateId}:
    patch:
      tags: [Sales Invoices]
      summary: تعديل نسبة ضريبة
      operationId: updateTaxRate
      parameters:
      - { name: taxRateId, in: path, required: true, schema: { type: string, pattern: '^[1-9][0-9]*$' } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TaxRateInput' } } } }
      responses:
        '200': { description: نجاح, content: { application/json: { schema: { $ref: '#/components/schemas/TaxRate' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
      x-permission: tax_rates.manage
      security: [{ SessionCookie: [], CsrfToken: [] }]
  /reports/receivables-aging:
    get:
      tags: [Reports]
      summary: تقرير أرصدة العملاء وأعمار الديون بالعملة الأساسية
      operationId: getReceivablesAging
      parameters:
      - { name: asOf, in: query, required: true, schema: { type: string, format: date } }
      - { name: customerId, in: query, schema: { type: string, pattern: '^[1-9][0-9]*$' } }
      responses:
        '200': { description: نجاح, content: { application/json: { schema: { $ref: '#/components/schemas/ReceivablesAgingReport' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
      x-permission: reports.receivables.view
      security: [{ SessionCookie: [] }]
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: sid
      description: معرف جلسة عشوائي؛ القيمة الخام لا تُخزن في قاعدة البيانات.
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: رمز Synchronizer مرتبط بالجلسة ومطلوب لكل طلب يغير الحالة.
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
  responses:
    ValidationError:
      description: خطأ في التنسيق أو الحقول
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    NotFound:
      description: الكيان غير موجود داخل الشركة الحالية؛ لا يكشف وجوده في شركة أخرى
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: تعارض إصدار أو تزامن أو تكرار
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    BusinessRuleViolation:
      description: انتهاك قاعدة أعمال، مثل فترة مغلقة أو قيد غير متوازن
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    PaginationMeta:
      type: object
      required:
      - page
      - pageSize
      - totalItems
      - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    Problem:
      type: object
      required:
      - code
      - messageAr
      - requestId
      properties:
        code:
          type: string
          enum:
          - VALIDATION_ERROR
          - UNAUTHENTICATED
          - FORBIDDEN
          - NOT_FOUND
          - CONFLICT
          - BUSINESS_RULE_VIOLATION
          - RATE_LIMITED
          - INTERNAL_ERROR
        messageAr:
          type: string
        requestId:
          type: string
          format: uuid
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        details:
          type: object
          additionalProperties: true
          description: تفاصيل آمنة فقط؛ لا تتضمن SQL أو Stack Trace أو قيماً حساسة.
...
    SalesInvoiceLineInput:
      type: object
      additionalProperties: false
      required: [description, quantity, unitPrice, discountAmount, revenueAccountId]
      properties:
        description: { type: string, minLength: 1, maxLength: 500 }
        quantity: { type: string, pattern: '^(0|[1-9][0-9]{0,14})\.[0-9]{4}$' }
        unitPrice: { type: string, pattern: '^(0|[1-9][0-9]{0,14})\.[0-9]{4}$' }
        discountAmount: { type: string, pattern: '^(0|[1-9][0-9]{0,14})\.[0-9]{4}$' }
        revenueAccountId: { type: string, pattern: '^[1-9][0-9]*$' }
        costCenterId: { type: [string, 'null'], pattern: '^[1-9][0-9]*$' }
        taxRateId: { type: [string, 'null'], pattern: '^[1-9][0-9]*$' }
    SalesInvoiceCreateRequest:
      type: object
      additionalProperties: false
      required: [documentType, fiscalPeriodId, documentDate, dueDate, description, customerId, currencyId, exchangeRate, lines]
      properties:
        documentType: { type: string, enum: [SALES_INVOICE, SALES_CREDIT_NOTE] }
        fiscalPeriodId: { type: string, pattern: '^[1-9][0-9]*$' }
        documentDate: { type: string, format: date }
        dueDate: { type: string, format: date }
        description: { type: string, minLength: 1, maxLength: 500 }
        customerId: { type: string, pattern: '^[1-9][0-9]*$' }
        sourceInvoiceId: { type: [string, 'null'], pattern: '^[1-9][0-9]*$' }
        currencyId: { type: string, pattern: '^[1-9][0-9]*$' }
        exchangeRate: { type: string, pattern: '^[0-9]{1,11}\.[0-9]{8}$' }
        customerAddress: { type: [string, 'null'], maxLength: 500 }
        notes: { type: [string, 'null'], maxLength: 1000 }
        lines: { type: array, minItems: 1, maxItems: 200, items: { $ref: '#/components/schemas/SalesInvoiceLineInput' } }
    SalesInvoiceUpdateRequest:
      type: object
      required: [version]
      properties:
        version: { type: integer, minimum: 0 }
        documentType: { type: string, enum: [SALES_INVOICE, SALES_CREDIT_NOTE] }
        fiscalPeriodId: { type: string, pattern: '^[1-9][0-9]*$' }
        documentDate: { type: string, format: date }
        dueDate: { type: string, format: date }
        description: { type: string, minLength: 1, maxLength: 500 }
        customerId: { type: string, pattern: '^[1-9][0-9]*$' }
        sourceInvoiceId: { type: [string, 'null'], pattern: '^[1-9][0-9]*$' }
        currencyId: { type: string, pattern: '^[1-9][0-9]*$' }
        exchangeRate: { type: string, pattern: '^[0-9]{1,11}\.[0-9]{8}$' }
        lines: { type: array, minItems: 1, maxItems: 200, items: { $ref: '#/components/schemas/SalesInvoiceLineInput' } }
    SalesInvoice:
      type: object
      required: [id, document, customerId, currencyId, dueDate, subtotal, discountTotal, taxTotal, total, baseTotal, paidAmount, creditedAmount, outstandingAmount, settlementStatus, lines]
      properties:
        id: { type: string, pattern: '^[1-9][0-9]*$' }
        document: { $ref: '#/components/schemas/AccountingDocumentHeader' }
        customerId: { type: string, pattern: '^[1-9][0-9]*$' }
        sourceInvoiceId: { type: [string, 'null'], pattern: '^[1-9][0-9]*$' }
        arJournalLineId: { type: [string, 'null'], pattern: '^[1-9][0-9]*$' }
        currencyId: { type: string, pattern: '^[1-9][0-9]*$' }
        dueDate: { type: string, format: date }
        subtotal: { type: string }
        discountTotal: { type: string }
        taxTotal: { type: string }
        total: { type: string }
        baseTotal: { type: string }
        paidAmount: { type: string }
        creditedAmount: { type: string }
        outstandingAmount: { type: string }
        settlementStatus: { type: string, enum: [OPEN, PARTIAL, PAID] }
        lines: { type: array, items: { type: object } }
...
    TaxRateInput:
      type: object
      required: [code, nameAr, rate]
      properties:
        code: { type: string, minLength: 1, maxLength: 40 }
        nameAr: { type: string, minLength: 1, maxLength: 120 }
        rate: { type: string, pattern: '^(0|[1-9][0-9]{0,2})\.[0-9]{4}$' }
        outputTaxAccountId: { type: [string, 'null'], pattern: '^[1-9][0-9]*$' }
        isActive: { type: boolean }
    TaxRate:
      allOf:
      - $ref: '#/components/schemas/TaxRateInput'
      - type: object
        required: [id, isActive]
        properties:
          id: { type: string, pattern: '^[1-9][0-9]*$' }
...
    ReceivablesAgingReport:
      type: object
      required: [asOf, baseCurrency, data, totals]
      properties:
        asOf: { type: string, format: date }
        baseCurrency: { type: object, required: [id, code, nameAr, decimals], properties: { id: { type: string }, code: { type: string }, nameAr: { type: string }, decimals: { type: integer } } }
        data: { type: array, items: { type: object } }
        totals: { type: object, properties: { current: { type: string }, days1To30: { type: string }, days31To60: { type: string }, days61To90: { type: string }, daysOver90: { type: string }, total: { type: string } } }
...
    AccountingDocumentStatus:
      type: string
      enum:
      - DRAFT
      - POSTED
      - CANCELLED
      - REVERSED
    AccountingDocumentType:
      type: string
      enum:
      - MANUAL_JOURNAL
      - RECEIPT
      - PAYMENT
      - SALES_INVOICE
      - SALES_CREDIT_NOTE
      - PURCHASE_INVOICE
      - PURCHASE_DEBIT_NOTE
      - PERIOD_CLOSE
    AccountingDocumentHeader:
      type: object
      required:
      - id
      - documentType
      - documentNumber
      - documentDate
      - description
      - status
      - fiscalPeriodId
      - version
      - createdBy
      - createdAt
      properties:
        id:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        documentType:
          $ref: '#/components/schemas/AccountingDocumentType'
        documentNumber:
          type: string
          maxLength: 60
        documentDate:
          type: string
          format: date
          example: '2026-07-30'
        description:
          type: string
          maxLength: 500
        status:
          $ref: '#/components/schemas/AccountingDocumentStatus'
        fiscalPeriodId:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        version:
          type: integer
          minimum: 0
          example: 0
        createdBy:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        createdAt:
          type: string
          format: date-time
          example: '2026-07-30T06:00:00.000Z'
        postedBy:
          oneOf:
          - type: string
            pattern: ^[1-9][0-9]*$
            example: '1001'
            description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
          - type: 'null'
        postedAt:
          oneOf:
          - type: string
            format: date-time
            example: '2026-07-30T06:00:00.000Z'
          - type: 'null'
        reversedByDocumentId:
          oneOf:
          - type: string
            pattern: ^[1-9][0-9]*$
            example: '1001'
            description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
          - type: 'null'
...